+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - marketplace list command
 - marketplace get command
+- environment list command
 
 ## [0.7.0] - 2023-06-26