+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - marketplace list command
 - marketplace get command
 - environment list command
+- pod list command
 
 ## [0.7.0] - 2023-06-26