+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2023-10-10
+
 ### Changed
 
 - update go version to 1.21.1
@@ -112,7 +114,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - create cli sdk
 - create cli renderer
 
-[unreleased]: https://github.com/mia-platform/miactl/compare/v0.7.0...HEAD
+[unreleased]: https://github.com/mia-platform/miactl/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/mia-platform/miactl/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/mia-platform/miactl/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/mia-platform/miactl/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/mia-platform/miactl/compare/v0.5.0...v0.6.0